Allow SlotFill for list items and cards
